Timed checkout
Checkout now takes a certain amount of time to complete.
Added a CustomerBusyState to Customer to determine whether the Customer
is INTERACTING/WAITING/NONE. This is similar to CustomerState, but this
is used for keeping track of the Customer's current activity.
Also shuffled around Status and StatBars so making the Status
automatically creates the StatBar that can be attached to an ActorExtra.