Shopping lists step :one:
Started the transition to having a Customer buy multiple items at
once!!!
Customers now have a Queue holding the items they want to buy. At the
moment it's just initialised to have one item but pretty sure I'll be
adding random items to these lists.
Also, added a Misc class with something to roll a number of ints that
are guaranteed to be different.